feat: :sparkles: Polymorphic types and theme types reoganization
