+++ libs/dana/src/foundations/theme/index.ts
@@ -5,3 +5,5 @@ export * from './themeProvider';
 export { useThemeContext };
 
 export * from './defaultTheme';
+
+export * from './types';
