+++ libs/dana/src/foundations/theme/types/Theme.ts
@@ -1,0 +1,11 @@
+import { Palette } from './Palette';
+
+export interface Theme {
+    name: string;
+    palette: Palette;
+    header?: any;
+    label?: any;
+    sideNav?: any;
+    text: string;
+    title: any;
+}
